modified REST endpoints to split up YAML and JSON, for easy UI use.
/services/metadata/{ID} always returns JSON
/services/metdata/yaml/{ID} always returns YAML
/services/metadata/autopopulate?repo={URL} returns JSON
/services/metadata/autopopulate/yaml?repo={URL} returns YAML

modified documentation README with this information.

Added "primary award" concept and field to sponsoring organizations.  This
will be the required award number to be validated against; any supplemental
award numbers may be supplied as the usual funding identifiers set.  This is
always assumed to be an "AwardNumber" type.

Modified unit tests and test case metadata.yml to accommodate this new
field and test its values.

Signed-off-by: Neal Ensor <ensorn@osti.gov>