modified DOI registration processes, if applicable:
if DataCite is configured, any SUBMIT or PUBLISH attempts will check DOI
registration for the passed-in values.  If the DOI prefix matches the DOECode
known prefix (in doecode.properties:${datacite.prefix}), a check will be made
to determine if the record being updated has already registered this DOI.

if not, registration entry will be made; if the DOI is registered to another
project, the registration will fail.

Should a DOI not be supplied, or one with an unknown prefix, DataCite
will NOT be notified, and will not attempt to post metadata information. Also,
if DataCite properties are not configured, it is assumed that any DOI has
been registered externally, and will not be validated nor will metadata
information be published to DataCite.

Metadata had unused DOI_STATUS field removed, as this serves as a replacement
procedure.

Note that the DOI value check is case-insensitive when checking for registered
DOI values.

Signed-off-by: Neal Ensor <ensorn@osti.gov>